# Conflicts:
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/monthlyattditem/FormCanUsedForTime.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/monthlyattendanceitem/MonthlyAttendanceItemUsedRepository.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/outsideot/JpaOutsideOTSettingRepository.java

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/outsideot/breakdown/KshstOutsideOtBrdPK.java
@@ -45,7 +45,7 @@ public class KshstOutsideOtBrdPK implements Serializable {
      * @param cid the cid
      * @param brdItemNo the brd item no
      */
-    public KshstOutsideOtBrdPK(String cid, short brdItemNo) {
+    public KshstOutsideOtBrdPK(String cid, int brdItemNo) {
         this.cid = cid;
         this.brdItemNo = brdItemNo;
     }
